refactor: replace previously implemented Button component with new one
