+++ app/dashboard/customers/page.tsx
@@ -1,5 +1,5 @@
 import { fetchCustomers } from '@/app/lib/data';
-import Button from '@/components/NextButton/Button';
+import Button from '@/components/Button/Button';
 import Link from 'next/link';
 import type { Metadata } from 'next';
